Drop the pdf field from lander.yaml templates
We don't actually use the "pdf" configuration in practice because the
reusable workflow
(SPHEREx/spherex-doc-workflows/.github/workflows/build-tex.yaml) renames
the PDF with the commit or tag name. This exposes a bug in the new
lander where the --pdf command line argument wasn't overriding the
lander.yaml settings; this will be fixed, but removing it from the
template is also good for long-term simplicity.