Add Markdown template for SSDC-TN
This is a template for writing Markdown documents in the SSDC-TN
document category. This uses the technote configuration in
spherex-sphinx added via https://github.com/SPHEREx/spherex-sphinx/pull/18

Similar templates for reStructuredText and even ipynb are to follow and
would be very similar to this template.

This template makes use of a new reusable workflow for building the HTML
site via the [tox](https://tox.wiki/) configuration. That workflow is
being PR'd as https://github.com/SPHEREx/spherex-doc-workflows/pull/10